Moving serializers to serializers folder
