+++ src/JWT/JWT.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using JWT.Serializers;
 
 namespace JWT
 {
